remote static from function declaration in source
static on functions means the symbols are not visible in any other
compilation unit.  this means they can be used in headers but not source
files (unless you include source files in other source files, which is a
bad practice).

see http://stackoverflow.com/questions/558122/what-is-a-static-function
for details.

there is a conflicting semantic between GCC's behavior and ISO C99 and
later, such that this code may work with GCC or C89 mode, but is likely
to break once compilers that default to C11 (e.g. recent GCC and Clang)
are used.